Improve logic.

+++ build.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-mono --runtime=v4.0 tools/NuGet/nuget.exe install FAKE -Version 4.4.2
+if [ ! -f /packages/FAKE.4.4.2/tools/FAKE.exe ]; then
+ mono --runtime=v4.0 tools/NuGet/nuget.exe install FAKE -Version 4.4.2
+fi
 mono --runtime=v4.0 packages/FAKE.4.4.2/tools/FAKE.exe build.fsx $@
